Task 6

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	float f;
 
 	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %i byte(s)\n", (unsigned int)sizeof(i));
-	printf("Size of a long int: %li byte(s)\n", (unsigned long)sizeof(li));
-	printf("Size of a long long int: %li byte(s)\n",
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(li));
+	printf("Size of a long long int: %llu byte(s)\n",
 	       (unsigned long long)sizeof(lli));
-	printf("Size of a float: %f byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 
 	return (0);
 }
